bot/requirements.txt
langgraph>=0.2.0
langchain-openai>=0.2.0
python-telegram-bot[webhooks]>=21.0
python-dotenv>=1.0.0
